moved check form into feature component
